requirements.txt
aiogram==3.1.1
asyncpg==0.29.0
apscheduler==3.10.4
python-dotenv==1.0.0
aiohttp==3.9.1
cryptography==41.0.7

